+++ app/src/main/java/app/hillforts/models/UnifiedJSONStrore.kt
@@ -72,6 +72,7 @@ class UnifiedJSONStore : UnifiedStore, AnkoLogger {
             foundHillfort.title = hillfort.title
             foundHillfort.description = hillfort.description
             foundHillfort.image = hillfort.image
+            foundHillfort.dateVisited = hillfort.dateVisited
             foundHillfort.lat = hillfort.lat
             foundHillfort.lng = hillfort.lng
             foundHillfort.zoom = hillfort.zoom